requirements.txt
pydub
requests
algoliasearch
pyTelegramBotAPI
google-cloud-speech
google-cloud-storage
